db_config.py
import mysql.connector

def get_db_connection():
    return mysql.connector.connect(
        host='localhost',
        user='root',
        password='thani123',
        database='candidate_assessment'
    )
